feat(*): Add rowDelete event and emit rowDeleted event #9994

+++ projects/igniteui-angular/src/lib/grids/grid-base.directive.ts
@@ -703,6 +703,19 @@ export abstract class IgxGridBaseDirective extends DisplayDensityBase implements
     public rowDeleted = new EventEmitter<IRowDataEventArgs>();
 
     /**
+     * Emmited when deleting a row
+     *
+     * @remarks
+     * Returns an `IRowDataEventArgs` object
+     * @example
+     * ```html
+     * <igx-grid #grid [data]="localData" (rowDelete)="rowDelete($event)" [height]="'305px'" [autoGenerate]="true"></igx-grid>
+     * ```
+     */
+    @Output()
+    public rowDelete = new EventEmitter<IGridEditEventArgs>();
+
+    /**
      * Emitted after column is resized.
      *
      * @remarks
@@ -4427,7 +4440,23 @@ export abstract class IgxGridBaseDirective extends DisplayDensityBase implements
 
     /** @hidden */
     public deleteRowById(rowId: any): any {
-        return this.gridAPI.deleteRowById(rowId);
+        const args = {
+            rowID: rowId,
+            cancel: false,
+            rowData: this.getRowData(rowId),
+            oldValue: null
+        };
+        this.rowDelete.emit(args);
+        if (args.cancel) {
+            return;
+        }
+
+        const record = this.gridAPI.deleteRowById(rowId);
+        if (record !== null && record !== undefined) {
+            //  TODO: should we emit this when cascadeOnDelete is true for each row?!?!
+            this.rowDeleted.emit({ data: record });
+        }
+        return record;
     }
 
     /**
